feat continuation: complete frontend logic to send invoice statements to medical aid providers
the logic for this feature has been completed. I am disabling the feature for the moment by removing the medicalAid email parameter from the fucntion repsosnible for sending the email statements in order to avoid sending false claims to medical aid schemes. I have partially completed the ui for it but it is still a work in progress